# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/btc/block/bestblockheight": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get best block height",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get best block height",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.BestBlockHeightResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/block/blockhash/{height}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get block hash with height",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get block hash with height",
                "parameters": [
                    {
                        "type": "string",
                        "description": "height",
                        "name": "height",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.BlockHashResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/block/{blockhash}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get raw block with blockhash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get raw block with blockhash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blockHash",
                        "name": "blockHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.RawBlockResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/rawtx/{txid}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get raw tx with txid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get raw tx with txid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "txid",
                        "name": "txid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.TxResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/tx": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "send Raw Transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "send Raw Transaction",
                "parameters": [
                    {
                        "description": "Signed transaction hex",
                        "name": "signedTxHex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Reject transactions whose fee rate is higher than the specified value, expressed in BTC/kB.default:",
                        "name": "maxfeerate",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.SendRawTxResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/tx/{txid}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get tx with txid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get tx with txid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "txid",
                        "name": "txid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.TxResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wire.BestBlockHeightResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "wire.BlockHashResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string",
                    "example": ""
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "wire.RawBlockResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string",
                    "example": ""
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "wire.SendRawTxResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string",
                    "example": "ae74538baa914f3799081ba78429d5d84f36a0127438e9f721dff584ac17b346"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "wire.TxResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}